+++ src/main/java-gen/dev/monogon/cue/lang/psi/CueAttribute.java
@@ -7,6 +7,15 @@ import com.intellij.psi.PsiElement;
 
 public interface CueAttribute extends CueCompositeElement {
 
+  @Nullable
+  CueAttrTokens getAttrTokens();
+
+  @Nullable
+  PsiElement getLeftParen();
+
+  @Nullable
+  PsiElement getRightParen();
+
   @NotNull PsiElement getAttributeNameElement();
 
   @NotNull String getAttributeName();