+++ src/main/java-gen/dev/monogon/cue/lang/psi/CueInterpolation.java
@@ -7,7 +7,7 @@ import com.intellij.psi.PsiElement;
 
 public interface CueInterpolation extends CueCompositeElement {
 
-  @NotNull
+  @Nullable
   CueExpression getExpression();
 
 }