+++ src/main/java/dev/monogon/cue/lang/psi/CueMultilineLiteral.java
@@ -1,0 +1,9 @@
+package dev.monogon.cue.lang.psi;
+
+import com.intellij.psi.PsiElement;
+
+public interface CueMultilineLiteral extends CueLiteral {
+    PsiElement getLiteralStartElement();
+
+    PsiElement getLiteralEndElement();
+}